moved cmip6_processor to processing
